requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fake-useragent==0.1.11
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
gensim==3.8.3
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
maxminddb==2.0.3
maxminddb-geolite2==2018.703
multidict==5.1.0
nltk==3.5
numpy==1.20.1
packaging==20.9
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
scipy==1.6.1
selenium==2.53.6
six==1.15.0
smart-open==4.2.0
soupsieve==2.2
tqdm==4.58.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.3
vcrpy==4.1.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
